+++ packages/google-cloud-dataplex/google/cloud/dataplex_v1/types/data_profile.py
@@ -81,10 +81,9 @@ class DataProfileSpec(proto.Message):
 
             Attributes:
                 results_table (str):
-                    Optional. The BigQuery table to export
-                    DataProfileScan results to. Format:
-
-                    projects/{project}/datasets/{dataset}/tables/{table}
+                    Optional. The BigQuery table to export DataProfileScan
+                    results to. Format:
+                    //bigquery.googleapis.com/projects/PROJECT_ID/datasets/DATASET_ID/tables/TABLE_ID
             """
 
             results_table: str = proto.Field(