+++ packages/google-cloud-functions/google/cloud/functions_v1/services/cloud_functions_service/transports/grpc_asyncio.py
@@ -498,6 +498,7 @@ class CloudFunctionsServiceGrpcAsyncIOTransport(CloudFunctionsServiceTransport):
         limited period and must be used within minutes after
         generation.
         For more information about the signed URL usage, see:
+
         https://cloud.google.com/storage/docs/access-control/signed-urls
 
         Returns: