+++ packages/google-cloud-netapp/google/cloud/netapp_v1/services/net_app/client.py
@@ -633,7 +633,8 @@ class NetAppClient(metaclass=NetAppClientMeta):
 
         Args:
             request (Union[google.cloud.netapp_v1.types.ListStoragePoolsRequest, dict]):
-                The request object.
+                The request object. ListStoragePoolsRequest lists Storage
+                Pools.
             parent (str):
                 Required. Parent value
                 This corresponds to the ``parent`` field
@@ -647,6 +648,8 @@ class NetAppClient(metaclass=NetAppClientMeta):
 
         Returns:
             google.cloud.netapp_v1.services.net_app.pagers.ListStoragePoolsPager:
+                ListStoragePoolsResponse is the
+                response to a ListStoragePoolsRequest.
                 Iterating over this object will yield
                 results and resolve additional pages
                 automatically.
@@ -757,7 +760,8 @@ class NetAppClient(metaclass=NetAppClientMeta):
 
         Args:
             request (Union[google.cloud.netapp_v1.types.CreateStoragePoolRequest, dict]):
-                The request object.
+                The request object. CreateStoragePoolRequest creates a
+                Storage Pool.
             parent (str):
                 Required. Value for parent.
                 This corresponds to the ``parent`` field
@@ -888,7 +892,8 @@ class NetAppClient(metaclass=NetAppClientMeta):
 
         Args:
             request (Union[google.cloud.netapp_v1.types.GetStoragePoolRequest, dict]):
-                The request object.
+                The request object. GetStoragePoolRequest gets a Storage
+                Pool.
             name (str):
                 Required. Name of the storage pool
                 This corresponds to the ``name`` field
@@ -1004,7 +1009,8 @@ class NetAppClient(metaclass=NetAppClientMeta):
 
         Args:
             request (Union[google.cloud.netapp_v1.types.UpdateStoragePoolRequest, dict]):
-                The request object.
+                The request object. UpdateStoragePoolRequest updates a
+                Storage Pool.
             storage_pool (google.cloud.netapp_v1.types.StoragePool):
                 Required. The pool being updated
                 This corresponds to the ``storage_pool`` field
@@ -1135,7 +1141,8 @@ class NetAppClient(metaclass=NetAppClientMeta):
 
         Args:
             request (Union[google.cloud.netapp_v1.types.DeleteStoragePoolRequest, dict]):
-                The request object.
+                The request object. DeleteStoragePoolRequest deletes a
+                Storage Pool.
             name (str):
                 Required. Name of the storage pool
                 This corresponds to the ``name`` field
@@ -2095,6 +2102,8 @@ class NetAppClient(metaclass=NetAppClientMeta):
 
         Returns:
             google.cloud.netapp_v1.types.Snapshot:
+                Snapshot is a point-in-time version
+                of a Volume's content.
 
         """
         # Create or coerce a protobuf request object.
@@ -2221,7 +2230,8 @@ class NetAppClient(metaclass=NetAppClientMeta):
                 An object representing a long-running operation.
 
                 The result type for the operation will be
-                :class:`google.cloud.netapp_v1.types.Snapshot`
+                :class:`google.cloud.netapp_v1.types.Snapshot` Snapshot
+                is a point-in-time version of a Volume's content.
 
         """
         # Create or coerce a protobuf request object.
@@ -2472,7 +2482,8 @@ class NetAppClient(metaclass=NetAppClientMeta):
                 An object representing a long-running operation.
 
                 The result type for the operation will be
-                :class:`google.cloud.netapp_v1.types.Snapshot`
+                :class:`google.cloud.netapp_v1.types.Snapshot` Snapshot
+                is a point-in-time version of a Volume's content.
 
         """
         # Create or coerce a protobuf request object.
@@ -3208,7 +3219,8 @@ class NetAppClient(metaclass=NetAppClientMeta):
 
         Args:
             request (Union[google.cloud.netapp_v1.types.ListKmsConfigsRequest, dict]):
-                The request object.
+                The request object. ListKmsConfigsRequest lists KMS
+                Configs.
             parent (str):
                 Required. Parent value
                 This corresponds to the ``parent`` field
@@ -3222,6 +3234,8 @@ class NetAppClient(metaclass=NetAppClientMeta):
 
         Returns:
             google.cloud.netapp_v1.services.net_app.pagers.ListKmsConfigsPager:
+                ListKmsConfigsResponse is the
+                response to a ListKmsConfigsRequest.
                 Iterating over this object will yield
                 results and resolve additional pages
                 automatically.
@@ -3328,7 +3342,8 @@ class NetAppClient(metaclass=NetAppClientMeta):
 
         Args:
             request (Union[google.cloud.netapp_v1.types.CreateKmsConfigRequest, dict]):
-                The request object.
+                The request object. CreateKmsConfigRequest creates a KMS
+                Config.
             parent (str):
                 Required. Value for parent.
                 This corresponds to the ``parent`` field
@@ -3459,7 +3474,8 @@ class NetAppClient(metaclass=NetAppClientMeta):
 
         Args:
             request (Union[google.cloud.netapp_v1.types.GetKmsConfigRequest, dict]):
-                The request object.
+                The request object. GetKmsConfigRequest gets a KMS
+                Config.
             name (str):
                 Required. Name of the KmsConfig
                 This corresponds to the ``name`` field
@@ -3566,7 +3582,8 @@ class NetAppClient(metaclass=NetAppClientMeta):
 
         Args:
             request (Union[google.cloud.netapp_v1.types.UpdateKmsConfigRequest, dict]):
-                The request object.
+                The request object. UpdateKmsConfigRequest updates a KMS
+                Config.
             kms_config (google.cloud.netapp_v1.types.KmsConfig):
                 Required. The KmsConfig being updated
                 This corresponds to the ``kms_config`` field
@@ -3877,7 +3894,8 @@ class NetAppClient(metaclass=NetAppClientMeta):
 
         Args:
             request (Union[google.cloud.netapp_v1.types.DeleteKmsConfigRequest, dict]):
-                The request object.
+                The request object. DeleteKmsConfigRequest deletes a KMS
+                Config.
             name (str):
                 Required. Name of the KmsConfig.
                 This corresponds to the ``name`` field
@@ -4884,6 +4902,331 @@ class NetAppClient(metaclass=NetAppClientMeta):
             and may cause errors in other clients!
         """
         self.transport.close()
+
+    def list_operations(
+        self,
+        request: Optional[operations_pb2.ListOperationsRequest] = None,
+        *,
+        retry: OptionalRetry = gapic_v1.method.DEFAULT,
+        timeout: Union[float, object] = gapic_v1.method.DEFAULT,
+        metadata: Sequence[Tuple[str, str]] = (),
+    ) -> operations_pb2.ListOperationsResponse:
+        r"""Lists operations that match the specified filter in the request.
+
+        Args:
+            request (:class:`~.operations_pb2.ListOperationsRequest`):
+                The request object. Request message for
+                `ListOperations` method.
+            retry (google.api_core.retry.Retry): Designation of what errors,
+                    if any, should be retried.
+            timeout (float): The timeout for this request.
+            metadata (Sequence[Tuple[str, str]]): Strings which should be
+                sent along with the request as metadata.
+        Returns:
+            ~.operations_pb2.ListOperationsResponse:
+                Response message for ``ListOperations`` method.
+        """
+        # Create or coerce a protobuf request object.
+        # The request isn't a proto-plus wrapped type,
+        # so it must be constructed via keyword expansion.
+        if isinstance(request, dict):
+            request = operations_pb2.ListOperationsRequest(**request)
+
+        # Wrap the RPC method; this adds retry and timeout information,
+        # and friendly error handling.
+        rpc = gapic_v1.method.wrap_method(
+            self._transport.list_operations,
+            default_timeout=None,
+            client_info=DEFAULT_CLIENT_INFO,
+        )
+
+        # Certain fields should be provided within the metadata header;
+        # add these here.
+        metadata = tuple(metadata) + (
+            gapic_v1.routing_header.to_grpc_metadata((("name", request.name),)),
+        )
+
+        # Send the request.
+        response = rpc(
+            request,
+            retry=retry,
+            timeout=timeout,
+            metadata=metadata,
+        )
+
+        # Done; return the response.
+        return response
+
+    def get_operation(
+        self,
+        request: Optional[operations_pb2.GetOperationRequest] = None,
+        *,
+        retry: OptionalRetry = gapic_v1.method.DEFAULT,
+        timeout: Union[float, object] = gapic_v1.method.DEFAULT,
+        metadata: Sequence[Tuple[str, str]] = (),
+    ) -> operations_pb2.Operation:
+        r"""Gets the latest state of a long-running operation.
+
+        Args:
+            request (:class:`~.operations_pb2.GetOperationRequest`):
+                The request object. Request message for
+                `GetOperation` method.
+            retry (google.api_core.retry.Retry): Designation of what errors,
+                    if any, should be retried.
+            timeout (float): The timeout for this request.
+            metadata (Sequence[Tuple[str, str]]): Strings which should be
+                sent along with the request as metadata.
+        Returns:
+            ~.operations_pb2.Operation:
+                An ``Operation`` object.
+        """
+        # Create or coerce a protobuf request object.
+        # The request isn't a proto-plus wrapped type,
+        # so it must be constructed via keyword expansion.
+        if isinstance(request, dict):
+            request = operations_pb2.GetOperationRequest(**request)
+
+        # Wrap the RPC method; this adds retry and timeout information,
+        # and friendly error handling.
+        rpc = gapic_v1.method.wrap_method(
+            self._transport.get_operation,
+            default_timeout=None,
+            client_info=DEFAULT_CLIENT_INFO,
+        )
+
+        # Certain fields should be provided within the metadata header;
+        # add these here.
+        metadata = tuple(metadata) + (
+            gapic_v1.routing_header.to_grpc_metadata((("name", request.name),)),
+        )
+
+        # Send the request.
+        response = rpc(
+            request,
+            retry=retry,
+            timeout=timeout,
+            metadata=metadata,
+        )
+
+        # Done; return the response.
+        return response
+
+    def delete_operation(
+        self,
+        request: Optional[operations_pb2.DeleteOperationRequest] = None,
+        *,
+        retry: OptionalRetry = gapic_v1.method.DEFAULT,
+        timeout: Union[float, object] = gapic_v1.method.DEFAULT,
+        metadata: Sequence[Tuple[str, str]] = (),
+    ) -> None:
+        r"""Deletes a long-running operation.
+
+        This method indicates that the client is no longer interested
+        in the operation result. It does not cancel the operation.
+        If the server doesn't support this method, it returns
+        `google.rpc.Code.UNIMPLEMENTED`.
+
+        Args:
+            request (:class:`~.operations_pb2.DeleteOperationRequest`):
+                The request object. Request message for
+                `DeleteOperation` method.
+            retry (google.api_core.retry.Retry): Designation of what errors,
+                    if any, should be retried.
+            timeout (float): The timeout for this request.
+            metadata (Sequence[Tuple[str, str]]): Strings which should be
+                sent along with the request as metadata.
+        Returns:
+            None
+        """
+        # Create or coerce a protobuf request object.
+        # The request isn't a proto-plus wrapped type,
+        # so it must be constructed via keyword expansion.
+        if isinstance(request, dict):
+            request = operations_pb2.DeleteOperationRequest(**request)
+
+        # Wrap the RPC method; this adds retry and timeout information,
+        # and friendly error handling.
+        rpc = gapic_v1.method.wrap_method(
+            self._transport.delete_operation,
+            default_timeout=None,
+            client_info=DEFAULT_CLIENT_INFO,
+        )
+
+        # Certain fields should be provided within the metadata header;
+        # add these here.
+        metadata = tuple(metadata) + (
+            gapic_v1.routing_header.to_grpc_metadata((("name", request.name),)),
+        )
+
+        # Send the request.
+        rpc(
+            request,
+            retry=retry,
+            timeout=timeout,
+            metadata=metadata,
+        )
+
+    def cancel_operation(
+        self,
+        request: Optional[operations_pb2.CancelOperationRequest] = None,
+        *,
+        retry: OptionalRetry = gapic_v1.method.DEFAULT,
+        timeout: Union[float, object] = gapic_v1.method.DEFAULT,
+        metadata: Sequence[Tuple[str, str]] = (),
+    ) -> None:
+        r"""Starts asynchronous cancellation on a long-running operation.
+
+        The server makes a best effort to cancel the operation, but success
+        is not guaranteed.  If the server doesn't support this method, it returns
+        `google.rpc.Code.UNIMPLEMENTED`.
+
+        Args:
+            request (:class:`~.operations_pb2.CancelOperationRequest`):
+                The request object. Request message for
+                `CancelOperation` method.
+            retry (google.api_core.retry.Retry): Designation of what errors,
+                    if any, should be retried.
+            timeout (float): The timeout for this request.
+            metadata (Sequence[Tuple[str, str]]): Strings which should be
+                sent along with the request as metadata.
+        Returns:
+            None
+        """
+        # Create or coerce a protobuf request object.
+        # The request isn't a proto-plus wrapped type,
+        # so it must be constructed via keyword expansion.
+        if isinstance(request, dict):
+            request = operations_pb2.CancelOperationRequest(**request)
+
+        # Wrap the RPC method; this adds retry and timeout information,
+        # and friendly error handling.
+        rpc = gapic_v1.method.wrap_method(
+            self._transport.cancel_operation,
+            default_timeout=None,
+            client_info=DEFAULT_CLIENT_INFO,
+        )
+
+        # Certain fields should be provided within the metadata header;
+        # add these here.
+        metadata = tuple(metadata) + (
+            gapic_v1.routing_header.to_grpc_metadata((("name", request.name),)),
+        )
+
+        # Send the request.
+        rpc(
+            request,
+            retry=retry,
+            timeout=timeout,
+            metadata=metadata,
+        )
+
+    def get_location(
+        self,
+        request: Optional[locations_pb2.GetLocationRequest] = None,
+        *,
+        retry: OptionalRetry = gapic_v1.method.DEFAULT,
+        timeout: Union[float, object] = gapic_v1.method.DEFAULT,
+        metadata: Sequence[Tuple[str, str]] = (),
+    ) -> locations_pb2.Location:
+        r"""Gets information about a location.
+
+        Args:
+            request (:class:`~.location_pb2.GetLocationRequest`):
+                The request object. Request message for
+                `GetLocation` method.
+            retry (google.api_core.retry.Retry): Designation of what errors,
+                 if any, should be retried.
+            timeout (float): The timeout for this request.
+            metadata (Sequence[Tuple[str, str]]): Strings which should be
+                sent along with the request as metadata.
+        Returns:
+            ~.location_pb2.Location:
+                Location object.
+        """
+        # Create or coerce a protobuf request object.
+        # The request isn't a proto-plus wrapped type,
+        # so it must be constructed via keyword expansion.
+        if isinstance(request, dict):
+            request = locations_pb2.GetLocationRequest(**request)
+
+        # Wrap the RPC method; this adds retry and timeout information,
+        # and friendly error handling.
+        rpc = gapic_v1.method.wrap_method(
+            self._transport.get_location,
+            default_timeout=None,
+            client_info=DEFAULT_CLIENT_INFO,
+        )
+
+        # Certain fields should be provided within the metadata header;
+        # add these here.
+        metadata = tuple(metadata) + (
+            gapic_v1.routing_header.to_grpc_metadata((("name", request.name),)),
+        )
+
+        # Send the request.
+        response = rpc(
+            request,
+            retry=retry,
+            timeout=timeout,
+            metadata=metadata,
+        )
+
+        # Done; return the response.
+        return response
+
+    def list_locations(
+        self,
+        request: Optional[locations_pb2.ListLocationsRequest] = None,
+        *,
+        retry: OptionalRetry = gapic_v1.method.DEFAULT,
+        timeout: Union[float, object] = gapic_v1.method.DEFAULT,
+        metadata: Sequence[Tuple[str, str]] = (),
+    ) -> locations_pb2.ListLocationsResponse:
+        r"""Lists information about the supported locations for this service.
+
+        Args:
+            request (:class:`~.location_pb2.ListLocationsRequest`):
+                The request object. Request message for
+                `ListLocations` method.
+            retry (google.api_core.retry.Retry): Designation of what errors,
+                 if any, should be retried.
+            timeout (float): The timeout for this request.
+            metadata (Sequence[Tuple[str, str]]): Strings which should be
+                sent along with the request as metadata.
+        Returns:
+            ~.location_pb2.ListLocationsResponse:
+                Response message for ``ListLocations`` method.
+        """
+        # Create or coerce a protobuf request object.
+        # The request isn't a proto-plus wrapped type,
+        # so it must be constructed via keyword expansion.
+        if isinstance(request, dict):
+            request = locations_pb2.ListLocationsRequest(**request)
+
+        # Wrap the RPC method; this adds retry and timeout information,
+        # and friendly error handling.
+        rpc = gapic_v1.method.wrap_method(
+            self._transport.list_locations,
+            default_timeout=None,
+            client_info=DEFAULT_CLIENT_INFO,
+        )
+
+        # Certain fields should be provided within the metadata header;
+        # add these here.
+        metadata = tuple(metadata) + (
+            gapic_v1.routing_header.to_grpc_metadata((("name", request.name),)),
+        )
+
+        # Send the request.
+        response = rpc(
+            request,
+            retry=retry,
+            timeout=timeout,
+            metadata=metadata,
+        )
+
+        # Done; return the response.
+        return response
 
 
 DEFAULT_CLIENT_INFO = gapic_v1.client_info.ClientInfo(