+++ packages/google-cloud-netapp/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [0.2.0](https://github.com/googleapis/google-cloud-python/compare/google-cloud-netapp-v0.1.0...google-cloud-netapp-v0.2.0) (2023-08-03)
+
+
+### Features
+
+* add initial version of NetApp v1 APIs ([de53f0e](https://github.com/googleapis/google-cloud-python/commit/de53f0efe25e71d0aa5b57b0989c4f0a491fa2ec))
+
+
+### Bug Fixes
+
+* remove netapp_v1beta1 client ([#11534](https://github.com/googleapis/google-cloud-python/issues/11534)) ([e7b16ed](https://github.com/googleapis/google-cloud-python/commit/e7b16ed17e06fa42858e5fdd35953805b8ca7e90))
+* update the default import to use `netapp_v1` ([de53f0e](https://github.com/googleapis/google-cloud-python/commit/de53f0efe25e71d0aa5b57b0989c4f0a491fa2ec))
+
 ## 0.1.0 (2023-07-20)